Move genesis related helpers from agave-snapshots to solana_genesis_utils
