+++ dev-bins/Cargo.toml
@@ -92,6 +92,7 @@ solana-feature-gate-interface = "3.0.0"
 solana-fee-calculator = "3.0.0"
 solana-genesis = { path = "../genesis", version = "=3.1.0", features = ["agave-unstable-api"] }
 solana-genesis-config = "3.0.0"
+solana-genesis-utils = { path = "../genesis-utils", version = "=3.1.0", features = ["agave-unstable-api"] }
 solana-geyser-plugin-manager = { path = "../geyser-plugin-manager", version = "=3.1.0", features = ["agave-unstable-api"] }
 solana-gossip = { path = "../gossip", version = "=3.1.0", features = ["agave-unstable-api"] }
 solana-hash = "3.0.0"
