+++ genesis/Cargo.toml
@@ -44,6 +44,7 @@ solana-epoch-schedule = "=3.0.0"
 solana-feature-gate-interface = "=3.0.0"
 solana-fee-calculator = "=3.0.0"
 solana-genesis-config = "=3.0.0"
+solana-genesis-utils = { workspace = true }
 solana-inflation = "=3.0.0"
 solana-keypair = "=3.0.1"
 solana-ledger = { workspace = true }
