+++ ledger-tool/Cargo.toml
@@ -55,6 +55,7 @@ solana-cost-model = { workspace = true }
 solana-entry = { workspace = true }
 solana-feature-gate-interface = { workspace = true }
 solana-genesis-config = { workspace = true }
+solana-genesis-utils = { workspace = true }
 solana-geyser-plugin-manager = { workspace = true }
 solana-gossip = { workspace = true }
 solana-hash = { workspace = true }
