+++ ledger/src/lib.rs
@@ -62,9 +62,9 @@ extern crate log;
 #[cfg(feature = "frozen-abi")]
 extern crate solana_frozen_abi_macro;
 
+mod wire_format_tests;
+
 #[doc(hidden)]
 pub mod macro_reexports {
-    pub use agave_snapshots::hardened_unpack::MAX_GENESIS_ARCHIVE_UNPACKED_SIZE;
+    pub use solana_genesis_utils::MAX_GENESIS_ARCHIVE_UNPACKED_SIZE;
 }
-
-mod wire_format_tests;
