+++ wen-restart/Cargo.toml
@@ -28,6 +28,7 @@ prost-types = { workspace = true }
 rayon = { workspace = true }
 solana-clock = { workspace = true }
 solana-entry = { workspace = true }
+solana-genesis-utils = { workspace = true }
 solana-gossip = { workspace = true }
 solana-hash = { workspace = true }
 solana-ledger = { workspace = true }
